+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyDataGet.java
@@ -7,6 +7,7 @@ import pnet.data.api.PnetDataClientException;
 import pnet.data.api.client.PnetDataClientResultPage;
 import pnet.data.api.util.AbstractGet;
 import pnet.data.api.util.ById;
+import pnet.data.api.util.ByMatchcode;
 import pnet.data.api.util.GetFunction;
 import pnet.data.api.util.Pair;
 import pnet.data.api.util.RestrictTenant;
@@ -15,8 +16,8 @@ import pnet.data.api.util.RestrictTenant;
  * @author cet
  *
  */
-public class CompanyDataGet extends AbstractGet<CompanyDataDTO, CompanyDataGet>
-    implements RestrictTenant<CompanyDataGet>, ById<CompanyDataDTO, CompanyDataGet>
+public class CompanyDataGet extends AbstractGet<CompanyDataDTO, CompanyDataGet> implements
+    RestrictTenant<CompanyDataGet>, ById<CompanyDataDTO, CompanyDataGet>, ByMatchcode<CompanyDataDTO, CompanyDataGet>
 {
 
     public CompanyDataGet(GetFunction<CompanyDataDTO> getFunction, List<Pair<String, Object>> restricts)